Refactor CertificatePinner into an interface and a Sha1CertificatePinner implementation such that different certificatepinners can be implemented.
