+++ okhttp-tests/src/test/java/com/squareup/okhttp/CertificatePinnerTest.java
@@ -54,17 +54,17 @@ public final class CertificatePinnerTest {
 
       keyPairA = sslContextBuilder.generateKeyPair();
       keypairACertificate1 = sslContextBuilder.selfSignedCertificate(keyPairA, "1");
-      keypairACertificate1Pin = CertificatePinner.pin(keypairACertificate1);
+      keypairACertificate1Pin = Sha1CertificatePinner.pin(keypairACertificate1);
       keypairACertificate1PinBase64 = pinToBase64(keypairACertificate1Pin);
 
       keyPairB = sslContextBuilder.generateKeyPair();
       keypairBCertificate1 = sslContextBuilder.selfSignedCertificate(keyPairB, "1");
-      keypairBCertificate1Pin = CertificatePinner.pin(keypairBCertificate1);
+      keypairBCertificate1Pin = Sha1CertificatePinner.pin(keypairBCertificate1);
       keypairBCertificate1PinBase64 = pinToBase64(keypairBCertificate1Pin);
 
       keyPairC = sslContextBuilder.generateKeyPair();
       keypairCCertificate1 = sslContextBuilder.selfSignedCertificate(keyPairC, "1");
-      keypairCCertificate1Pin = CertificatePinner.pin(keypairCCertificate1);
+      keypairCCertificate1Pin = Sha1CertificatePinner.pin(keypairCCertificate1);
     } catch (GeneralSecurityException e) {
       throw new AssertionError(e);
     }
@@ -75,7 +75,7 @@ public final class CertificatePinnerTest {
   }
 
   @Test public void malformedPin() throws Exception {
-    CertificatePinner.Builder builder = new CertificatePinner.Builder();
+    Sha1CertificatePinner.Builder builder = new Sha1CertificatePinner.Builder();
     try {
       builder.add("example.com", "md5/DmxUShsZuNiqPQsX2Oi9uv2sCnw=");
       fail();
@@ -84,7 +84,7 @@ public final class CertificatePinnerTest {
   }
 
   @Test public void malformedBase64() throws Exception {
-    CertificatePinner.Builder builder = new CertificatePinner.Builder();
+    Sha1CertificatePinner.Builder builder = new Sha1CertificatePinner.Builder();
     try {
       builder.add("example.com", "sha1/DmxUShsZuNiqPQsX2Oi9uv2sCnw*");
       fail();
@@ -95,10 +95,10 @@ public final class CertificatePinnerTest {
   /** Multiple certificates generated from the same keypair have the same pin. */
   @Test public void sameKeypairSamePin() throws Exception {
     X509Certificate keypairACertificate2 = sslContextBuilder.selfSignedCertificate(keyPairA, "2");
-    String keypairACertificate2Pin = CertificatePinner.pin(keypairACertificate2);
+    String keypairACertificate2Pin = Sha1CertificatePinner.pin(keypairACertificate2);
 
     X509Certificate keypairBCertificate2 = sslContextBuilder.selfSignedCertificate(keyPairB, "2");
-    String keypairBCertificate2Pin = CertificatePinner.pin(keypairBCertificate2);
+    String keypairBCertificate2Pin = Sha1CertificatePinner.pin(keypairBCertificate2);
 
     assertTrue(keypairACertificate1Pin.equals(keypairACertificate2Pin));
     assertTrue(keypairBCertificate1Pin.equals(keypairBCertificate2Pin));
@@ -106,7 +106,7 @@ public final class CertificatePinnerTest {
   }
 
   @Test public void successfulCheck() throws Exception {
-    CertificatePinner certificatePinner = new CertificatePinner.Builder()
+    Sha1CertificatePinner certificatePinner = new Sha1CertificatePinner.Builder()
         .add("example.com", keypairACertificate1Pin)
         .build();
 
@@ -114,7 +114,7 @@ public final class CertificatePinnerTest {
   }
 
   @Test public void successfulMatchAcceptsAnyMatchingCertificate() throws Exception {
-    CertificatePinner certificatePinner = new CertificatePinner.Builder()
+    Sha1CertificatePinner certificatePinner = new Sha1CertificatePinner.Builder()
         .add("example.com", keypairBCertificate1Pin)
         .build();
 
@@ -122,7 +122,7 @@ public final class CertificatePinnerTest {
   }
 
   @Test public void unsuccessfulCheck() throws Exception {
-    CertificatePinner certificatePinner = new CertificatePinner.Builder()
+    Sha1CertificatePinner certificatePinner = new Sha1CertificatePinner.Builder()
         .add("example.com", keypairACertificate1Pin)
         .build();
 
@@ -134,7 +134,7 @@ public final class CertificatePinnerTest {
   }
 
   @Test public void multipleCertificatesForOneHostname() throws Exception {
-    CertificatePinner certificatePinner = new CertificatePinner.Builder()
+    Sha1CertificatePinner certificatePinner = new Sha1CertificatePinner.Builder()
         .add("example.com", keypairACertificate1Pin, keypairBCertificate1Pin)
         .build();
 
@@ -143,7 +143,7 @@ public final class CertificatePinnerTest {
   }
 
   @Test public void multipleHostnamesForOneCertificate() throws Exception {
-    CertificatePinner certificatePinner = new CertificatePinner.Builder()
+    Sha1CertificatePinner certificatePinner = new Sha1CertificatePinner.Builder()
         .add("example.com", keypairACertificate1Pin)
         .add("www.example.com", keypairACertificate1Pin)
         .build();
@@ -153,12 +153,12 @@ public final class CertificatePinnerTest {
   }
 
   @Test public void absentHostnameMatches() throws Exception {
-    CertificatePinner certificatePinner = new CertificatePinner.Builder().build();
+    Sha1CertificatePinner certificatePinner = new Sha1CertificatePinner.Builder().build();
     certificatePinner.check("example.com", keypairACertificate1);
   }
 
   @Test public void successfulCheckForWildcardHostname() throws Exception {
-    CertificatePinner certificatePinner = new CertificatePinner.Builder()
+    Sha1CertificatePinner certificatePinner = new Sha1CertificatePinner.Builder()
         .add("*.example.com", keypairACertificate1Pin)
         .build();
 
@@ -166,7 +166,7 @@ public final class CertificatePinnerTest {
   }
 
   @Test public void successfulMatchAcceptsAnyMatchingCertificateForWildcardHostname() throws Exception {
-    CertificatePinner certificatePinner = new CertificatePinner.Builder()
+    Sha1CertificatePinner certificatePinner = new Sha1CertificatePinner.Builder()
         .add("*.example.com", keypairBCertificate1Pin)
         .build();
 
@@ -174,7 +174,7 @@ public final class CertificatePinnerTest {
   }
 
   @Test public void unsuccessfulCheckForWildcardHostname() throws Exception {
-    CertificatePinner certificatePinner = new CertificatePinner.Builder()
+    Sha1CertificatePinner certificatePinner = new Sha1CertificatePinner.Builder()
         .add("*.example.com", keypairACertificate1Pin)
         .build();
 
@@ -186,7 +186,7 @@ public final class CertificatePinnerTest {
   }
 
   @Test public void multipleCertificatesForOneWildcardHostname() throws Exception {
-    CertificatePinner certificatePinner = new CertificatePinner.Builder()
+    Sha1CertificatePinner certificatePinner = new Sha1CertificatePinner.Builder()
         .add("*.example.com", keypairACertificate1Pin, keypairBCertificate1Pin)
         .build();
 
@@ -195,7 +195,7 @@ public final class CertificatePinnerTest {
   }
 
   @Test public void successfulCheckForOneHostnameWithWildcardAndDirectCertificate() throws Exception {
-    CertificatePinner certificatePinner = new CertificatePinner.Builder()
+    Sha1CertificatePinner certificatePinner = new Sha1CertificatePinner.Builder()
         .add("*.example.com", keypairACertificate1Pin)
         .add("a.example.com", keypairBCertificate1Pin)
         .build();
@@ -205,7 +205,7 @@ public final class CertificatePinnerTest {
   }
 
   @Test public void unsuccessfulCheckForOneHostnameWithWildcardAndDirectCertificate() throws Exception {
-    CertificatePinner certificatePinner = new CertificatePinner.Builder()
+    Sha1CertificatePinner certificatePinner = new Sha1CertificatePinner.Builder()
         .add("*.example.com", keypairACertificate1Pin)
         .add("a.example.com", keypairBCertificate1Pin)
         .build();
@@ -218,7 +218,7 @@ public final class CertificatePinnerTest {
   }
 
   @Test public void successfulFindMatchingPins() {
-    CertificatePinner certificatePinner = new CertificatePinner.Builder()
+    Sha1CertificatePinner certificatePinner = new Sha1CertificatePinner.Builder()
         .add("first.com", keypairACertificate1Pin, keypairBCertificate1Pin)
         .add("second.com", keypairCCertificate1Pin)
         .build();
@@ -230,7 +230,7 @@ public final class CertificatePinnerTest {
   }
 
   @Test public void successfulFindMatchingPinsForWildcardAndDirectCertificates() {
-    CertificatePinner certificatePinner = new CertificatePinner.Builder()
+    Sha1CertificatePinner certificatePinner = new Sha1CertificatePinner.Builder()
         .add("*.example.com", keypairACertificate1Pin)
         .add("a.example.com", keypairBCertificate1Pin)
         .add("b.example.com", keypairCCertificate1Pin)
@@ -243,7 +243,7 @@ public final class CertificatePinnerTest {
   }
 
   @Test public void wildcardHostnameShouldNotMatchThroughDot() throws Exception {
-    CertificatePinner certificatePinner = new CertificatePinner.Builder()
+    Sha1CertificatePinner certificatePinner = new Sha1CertificatePinner.Builder()
         .add("*.example.com", keypairACertificate1Pin)
         .build();
