+++ okhttp-tests/src/test/java/com/squareup/okhttp/OkHttpClientTest.java
@@ -115,7 +115,7 @@ public final class OkHttpClientTest {
     Authenticator authenticator = new RecordingAuthenticator();
     SocketFactory socketFactory = SocketFactory.getDefault(); // Global isn't configurable.
     OkHostnameVerifier hostnameVerifier = OkHostnameVerifier.INSTANCE; // Global isn't configurable.
-    CertificatePinner certificatePinner = CertificatePinner.DEFAULT; // Global isn't configurable.
+    CertificatePinner certificatePinner = Sha1CertificatePinner.DEFAULT; // Global isn't configurable.
 
     CookieManager.setDefault(cookieManager);
     ProxySelector.setDefault(proxySelector);
